+++ src/mcmu/__main__.py
@@ -1,7 +1,6 @@
 """Main helper script (python3 -m mcmu)"""
 
 from .cli import CLI
-from .shared import logger
 
 
 def main() -> int:
@@ -10,10 +9,9 @@ def main() -> int:
         cli = CLI()  # Set up CLI class
         return cli.cli()  # Run CLI
     except KeyboardInterrupt:  # If CTRL+C
-        print("^C pressed. Exiting...")  # Print error
+        print(" pressed, exiting...")  # Print error
         return 130  # Exit
 
 
 if __name__ == "__main__":
-    logger.setLevel(10)  # Set to debug
     raise SystemExit(main())