+++ src/mcmu/shared.py
@@ -1,14 +1,61 @@
 """Shared helper scripts for MCMU"""
 
-from os import listdir, getenv, environ
+from logging import getLogger
+from os import environ, getenv, listdir, replace
 from pathlib import Path
 from platform import system
 from re import match
-from . import logger
 from .api import ModrinthAPI
-from .mods import Mod
 
+logger = getLogger(__name__)
 ModAPI = ModrinthAPI()
+
+
+class Mod:
+    """Mod class"""
+    def __init__(
+        self,
+        name: str,
+        version: str,
+        file_name: Path,
+        mod_folder: Path
+    ):
+        self.name = name
+        self.version = version
+        self.file_name = file_name
+        self.mod_folder = mod_folder
+        self.file = Path(mod_folder, file_name)
+        self.enabled = True
+        if file_name.endswith(".disabled"):
+            self.enabled = False
+
+    def enable(self):
+        """Enable the mod"""
+        if not self.enabled:
+            replace(
+                Path(
+                    self.mod_folder,
+                    f"{self.name}_version_{self.version}.jar.disabled"
+                ),
+                Path(
+                    self.mod_folder,
+                    f"{self.name}_version_{self.version}.jar"
+                )
+            )
+
+    def disable(self):
+        """Disable mod"""
+        if self.enabled:
+            replace(
+                Path(
+                    self.mod_folder,
+                    f"{self.name}_version_{self.version}.jar"
+                ),
+                Path(
+                    self.mod_folder,
+                    f"{self.name}_version_{self.version}.jar.disabled"
+                )
+            )
 
 
 def get_categories() -> list:
@@ -198,7 +245,7 @@ def update_mods(
     channel: list = "release"
 ) -> bool:
     """Updates mods"""
-    for mod_name in mods.items():
+    for mod_name in mods:
         latest_version = get_latest_version(
             mod_name,
             mod_loader,